`curve_analysis/data_processing.py` -> `utils.py`
