+++ docs/tutorials/getting_started.rst
@@ -7,8 +7,8 @@ Installation
 
 Qiskit Experiments is built on top of Qiskit, so we recommend that you first install
 Qiskit following its :external+qiskit:doc:`installation guide <getting_started>`. Qiskit
-Experiments supports the same platforms and Python versions (currently 3.7+) as Qiskit
-itself.
+Experiments supports the same platforms and Python versions (currently **3.7+**) as
+Qiskit itself.
 
 Qiskit Experiments releases can be installed via the Python package manager ``pip``:
